Add glooctl commands for auth config, restructure virtual service glooctl commands
