+++ docs/content/cli/glooctl_get.md
@@ -40,6 +40,7 @@ glooctl get [flags]
 ### SEE ALSO
 
 * [glooctl](../glooctl)	 - CLI for Gloo
+* [glooctl get authconfig](../glooctl_get_authconfig)	 - read an authconfig or list authconfigs in a namespace
 * [glooctl get proxy](../glooctl_get_proxy)	 - read a proxy or list proxies in a namespace
 * [glooctl get routetable](../glooctl_get_routetable)	 - read a route table or list route tables in a namespace
 * [glooctl get upstream](../glooctl_get_upstream)	 - read an upstream or list upstreams in a namespace
