+++ docs/content/cli/glooctl_get_authconfig.md
@@ -1,0 +1,43 @@
+---
+title: "glooctl get authconfig"
+weight: 5
+---
+## glooctl get authconfig
+
+read an authconfig or list authconfigs in a namespace
+
+### Synopsis
+
+usage: glooctl get authconfig [NAME] [--namespace=namespace] [-o FORMAT]
+
+```
+glooctl get authconfig [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for authconfig
+```
+
+### Options inherited from parent commands
+
+```
+  -c, --config string              set the path to the glooctl config file (default "<home_directory>/.gloo/glooctl-config.yaml")
+      --consul-address string      address of the Consul server. Use with --use-consul (default "127.0.0.1:8500")
+      --consul-datacenter string   Datacenter to use. If not provided, the default agent datacenter is used. Use with --use-consul
+      --consul-root-key string     key prefix for for Consul key-value storage. (default "gloo")
+      --consul-scheme string       URI scheme for the Consul server. Use with --use-consul (default "http")
+      --consul-token string        Token is used to provide a per-request ACL token which overrides the agent's default token. Use with --use-consul
+  -i, --interactive                use interactive mode
+      --kubeconfig string          kubeconfig to use, if not standard one
+      --name string                name of the resource to read or write
+  -n, --namespace string           namespace for reading or writing resources (default "gloo-system")
+  -o, --output OutputType          output format: (yaml, json, table, kube-yaml, wide) (default table)
+      --use-consul                 use Consul Key-Value storage as the backend for reading and writing config (VirtualServices, Upstreams, and Proxies)
+```
+
+### SEE ALSO
+
+* [glooctl get](../glooctl_get)	 - Display one or a list of Gloo resources
+
