+++ projects/gloo/cli/pkg/cmd/create/authconfig/authconfig_suite_test.go
@@ -1,0 +1,13 @@
+package authconfig_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestAuthConfig(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "AuthConfig Suite")
+}
